requirements/base.txt
Django==3.0.7
Channels==2.3.1
channels-redis==2.4.1
dj-database-url==0.5.0
whitenoise==4.1.4
bootstrap4==0.1.0
celery==4.3.0
django-pandas==0.6.1
django-redis==4.11.0
django-redis-cache==2.1.0
django-select-multiple-field==0.4.2
django-select2==7.1.1
numpy~=1.18.4
openpyxl~=3.0.3
pandas~=1.0.3
redis==3.3.11
redis-cache==0.1.5
requests
xlrd
xlutils
xlwt

asgiref~=3.2.7
djangorestframework~=3.11.0
django-cors-headers==3.3.0
djangorestframework-xml==2.0.0
